refactor(bezier-codemod): separate interpolation transform
